Move instance convertors into their own namespace for consistency
